fixed: broken detection whether installer should be used
this previously determined whether an installer is supported based on the active game but then the installation was (correctly) using the game for which the mod was being installed.

Therefore when installing a mod for a different game, if the active game was supported by bepinex but the mod was installed for a different game, this would cause an exception.
If the active game wasn't supported by bepinex but the target game and the mod was, then the installer wasn't being used even though it should be

fixes Nexus-Mods/Vortex#12935